+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - New setting - `platformSettings.condition`.
+- Added `platformSettings.platform[platform].remove`.
 
 ### Modified